extract filterLines to new text module
